requirements.txt
streamlit==1.28.0
anthropic==0.34.0
pandas==2.0.3
python-dotenv==1.0.0
sentence-transformers==2.7.0
huggingface-hub==0.23.0
faiss-cpu==1.7.4
numpy==1.24.3
torch==2.0.1
PyPDF2==3.0.1
python-docx==1.1.0